requirements.txt
# General libraries

pillow>=11.2.1